feat: postgres auto-provisioning, secrets hygiene, and CI-green programs
- Backend: add shared db/bootstrap (SSL-aware pool honoring sslmode, idempotent
  migrations + seed-if-empty), auto-run on startup for the postgres driver
  (DB_AUTO_MIGRATE / DB_AUTO_SEED), reuse it in scripts/migrate + seed
- Repos: fix JSONB params parsing and strategy seed nulls; expand Postgres
  integration tests to full CRUD/lifecycle coverage; run them in CI
- Secrets: untrack deployer/oracle keypairs and gitignore them; sanitize leaked
  Helius API key from devnet-config; document rotation in SECURITY/RUNBOOK
- Programs: declare Anchor cfg values via [lints.rust] check-cfg and fix real
  clippy lints so `cargo clippy --workspace -- -D warnings` is green

+++ programs/treasury/src/instructions/mod.rs
@@ -245,7 +245,7 @@ pub fn withdraw_revenue_handler(ctx: Context<WithdrawRevenue>, amount: u64) -> R
         (spent as u128)
             .checked_add(amount as u128)
             .ok_or(TreasuryError::MathOverflow)?
-            <= cap as u128,
+            <= cap,
         TreasuryError::WithdrawCapExceeded
     );
 
@@ -315,6 +315,7 @@ pub struct UpdateConfig<'info> {    #[account(
     pub governance: Signer<'info>,
 }
 
+#[allow(clippy::too_many_arguments)]
 pub fn update_config_handler(
     ctx: Context<UpdateConfig>,
     buyback_authority: Option<Pubkey>,
